refactor: split file write functions
